Changes required after review
