+++ public/swagger.yaml
@@ -49,7 +49,7 @@ paths:
               schema:
                 $ref: '#/components/schemas/ResponseBody'
             
-  /secured:
+  /securedPing:
     get:
       operationId: getSecured
       tags:
