+++ server/helpers/auth.js
@@ -21,13 +21,39 @@ const util = require('../helpers/util');
 /**
  * Set up logging
  */
-const logger = log4js.getLogger('middlewares - auth');
+const logger = log4js.getLogger('helpers - auth');
 logger.setLevel(config.logLevel);
 
 
 logger.debug('importing up app id config');
 
-function getAudience(token) {
+/**
+ * auth object
+ */
+const auth = {};
+
+/**
+* filter audience, or list of client ID's, that can access this
+*/
+auth.filter = whitelist => (req, res, next) => {
+  const accessTokenString = auth.getAccessToken(req, next);
+  const accessTokenPayload = jwt.decode(accessTokenString);
+
+  const audience = auth.getAudience(accessTokenPayload);
+
+  if (whitelist.includes(audience)) {
+    logger.debug('client is in whitelist, proceed');
+    next();
+  } else {
+    logger.debug('token does not have the appropriate access rights (aud)');
+    res.status(401).send({ error: 'invalid_grant', message: 'This token does not have the appropriate access rights (aud)' });
+  }
+};
+
+/**
+ * Takes the token and determines what version to return formatted audience
+ */
+auth.getAudience = (token) => {
   logger.info(`token aud: ${token.aud}`);
   let audience;
   if (Array.isArray(token.aud)) {
@@ -39,29 +65,25 @@ function getAudience(token) {
   }
   logger.debug(`audience: ${audience}`);
   return audience;
-}
+};
 
 /**
- * auth object
+ * Helper to grab auth header from request
  */
-const auth = {};
-
-/**
-* filter audience, or list of client ID's, that can access this
-*/
-auth.filter = whitelist => (req, res, next) => {
-  const accessTokenString = util.getAccessToken(req, next);
-  const accessTokenPayload = jwt.decode(accessTokenString);
-
-  const audience = getAudience(accessTokenPayload);
-
-  if (whitelist.includes(audience)) {
-    logger.debug('client is in whitelist, proceed');
-    next();
-  } else {
-    logger.debug('token does not have the appropriate access rights (aud)');
-    res.status(401).send({ error: 'invalid_grant', message: 'This token does not have the appropriate access rights (aud)' });
+auth.getAccessToken = (req, next) => {
+  const authHeader = req.headers.authorization;
+  if (!authHeader) {
+    logger.debug('Authorization header not found');
+    return next(new Error('Authorization header not found'));
   }
+
+  // Check for Bearer type in header component
+  const authHeaderComponents = authHeader.split(' ');
+  if (authHeaderComponents[0].indexOf('Bearer') !== 0) {
+    logger.debug('Malformed authorization header');
+    return next(new Error('Malformed authorization header'));
+  }
+  return authHeaderComponents[1];
 };
 
 module.exports = auth;
